AEZ-core: quit early on `Cy` block tag mismatch.
The scheme hashes the first pass encryption and mixes it into
the last block (`Cy`). It's possible to decrypt the last block
before continuing with the rest of the message. This allows us
to check the last 16 bytes of the tag early. (Note that the
whole tag should still be checked in decryption.)